changed class structure in Dataset.py
